feat(automations): adds morning message and workday timer
additional sensors were created to translate the weather and predicted weather conditions. An
automation was added to send a "morning message" with the current and predicted weather conditions,
including travel time to work. The other added automation checks whether I enter the work zone, and
starts a countdown timer. At the end of the timer (= end of workday) a notification is send to my
phone to inform me that the workday is over, and what the current travel time to home is based on
the Waze add-on